+++ lib/src/formatters/cartao_bancario_input_formatter.dart
@@ -7,31 +7,31 @@ class CartaoBancarioInputFormatter extends TextInputFormatter {
 
   @override
   TextEditingValue formatEditUpdate(
-      TextEditingValue oldValue, TextEditingValue newValue) {
-    final newTextLength = newValue.text.length;
-    var selectionIndex = newValue.selection.end;
+      TextEditingValue valorAntigo, TextEditingValue valorNovo) {
+    final novoTextLength = valorNovo.text.length;
+    var selectionIndex = valorNovo.selection.end;
 
-    if (newTextLength > maxLength) {
-      return oldValue;
+    if (novoTextLength > maxLength) {
+      return valorAntigo;
     }
 
     var usedSubstringIndex = 0;
     final newText = StringBuffer();
 
-    if (newTextLength >= 4) {
-      newText.write(newValue.text.substring(0, usedSubstringIndex = 4) + ' ');
-      if (newValue.selection.end >= 5) selectionIndex++;
+    if (novoTextLength >= 4) {
+      newText.write(valorNovo.text.substring(0, usedSubstringIndex = 4) + ' ');
+      if (valorNovo.selection.end >= 5) selectionIndex++;
     }
-    if (newTextLength >= 8) {
-      newText.write(newValue.text.substring(4, usedSubstringIndex = 8) + ' ');
-      if (newValue.selection.end >= 9) selectionIndex++;
+    if (novoTextLength >= 8) {
+      newText.write(valorNovo.text.substring(4, usedSubstringIndex = 8) + ' ');
+      if (valorNovo.selection.end >= 9) selectionIndex++;
     }
-    if (newTextLength >= 12) {
-      newText.write(newValue.text.substring(8, usedSubstringIndex = 12) + ' ');
-      if (newValue.selection.end >= 13) selectionIndex++;
+    if (novoTextLength >= 12) {
+      newText.write(valorNovo.text.substring(8, usedSubstringIndex = 12) + ' ');
+      if (valorNovo.selection.end >= 13) selectionIndex++;
     }
-    if (newTextLength >= usedSubstringIndex) {
-      newText.write(newValue.text.substring(usedSubstringIndex));
+    if (novoTextLength >= usedSubstringIndex) {
+      newText.write(valorNovo.text.substring(usedSubstringIndex));
     }
 
     return TextEditingValue(