+++ lib/src/formatters/cnpj_input_formatter.dart
@@ -10,35 +10,35 @@ class CnpjInputFormatter extends TextInputFormatter
 
   @override
   TextEditingValue formatEditUpdate(
-      TextEditingValue oldValue, TextEditingValue newValue) {
-    final newTextLength = newValue.text.length;
-    var selectionIndex = newValue.selection.end;
+      TextEditingValue valorAntigo, TextEditingValue valorNovo) {
+    final novoTextLength = valorNovo.text.length;
+    var selectionIndex = valorNovo.selection.end;
 
-    if (newTextLength > maxLength) {
-      return oldValue;
+    if (novoTextLength > maxLength) {
+      return valorAntigo;
     }
 
     var usedSubstringIndex = 0;
     final newText = StringBuffer();
 
-    if (newTextLength >= 3) {
-      newText.write(newValue.text.substring(0, usedSubstringIndex = 2) + '.');
-      if (newValue.selection.end >= 2) selectionIndex++;
+    if (novoTextLength >= 3) {
+      newText.write(valorNovo.text.substring(0, usedSubstringIndex = 2) + '.');
+      if (valorNovo.selection.end >= 2) selectionIndex++;
     }
-    if (newTextLength >= 6) {
-      newText.write(newValue.text.substring(2, usedSubstringIndex = 5) + '.');
-      if (newValue.selection.end >= 5) selectionIndex++;
+    if (novoTextLength >= 6) {
+      newText.write(valorNovo.text.substring(2, usedSubstringIndex = 5) + '.');
+      if (valorNovo.selection.end >= 5) selectionIndex++;
     }
-    if (newTextLength >= 9) {
-      newText.write(newValue.text.substring(5, usedSubstringIndex = 8) + '/');
-      if (newValue.selection.end >= 8) selectionIndex++;
+    if (novoTextLength >= 9) {
+      newText.write(valorNovo.text.substring(5, usedSubstringIndex = 8) + '/');
+      if (valorNovo.selection.end >= 8) selectionIndex++;
     }
-    if (newTextLength >= 13) {
-      newText.write(newValue.text.substring(8, usedSubstringIndex = 12) + '-');
-      if (newValue.selection.end >= 12) selectionIndex++;
+    if (novoTextLength >= 13) {
+      newText.write(valorNovo.text.substring(8, usedSubstringIndex = 12) + '-');
+      if (valorNovo.selection.end >= 12) selectionIndex++;
     }
-    if (newTextLength >= usedSubstringIndex) {
-      newText.write(newValue.text.substring(usedSubstringIndex));
+    if (novoTextLength >= usedSubstringIndex) {
+      newText.write(valorNovo.text.substring(usedSubstringIndex));
     }
 
     return TextEditingValue(