+++ lib/src/formatters/data_input_formatter.dart
@@ -7,27 +7,27 @@ class DataInputFormatter extends TextInputFormatter {
 
   @override
   TextEditingValue formatEditUpdate(
-      TextEditingValue oldValue, TextEditingValue newValue) {
-    final newTextLength = newValue.text.length;
-    var selectionIndex = newValue.selection.end;
+      TextEditingValue valorAntigo, TextEditingValue valorNovo) {
+    final novoTextLength = valorNovo.text.length;
+    var selectionIndex = valorNovo.selection.end;
 
-    if (newTextLength > maxLength) {
-      return oldValue;
+    if (novoTextLength > maxLength) {
+      return valorAntigo;
     }
 
     var usedSubstringIndex = 0;
     final newText = StringBuffer();
 
-    if (newTextLength >= 3) {
-      newText.write(newValue.text.substring(0, usedSubstringIndex = 2) + '/');
-      if (newValue.selection.end >= 2) selectionIndex++;
+    if (novoTextLength >= 3) {
+      newText.write(valorNovo.text.substring(0, usedSubstringIndex = 2) + '/');
+      if (valorNovo.selection.end >= 2) selectionIndex++;
     }
-    if (newTextLength >= 5) {
-      newText.write(newValue.text.substring(2, usedSubstringIndex = 4) + '/');
-      if (newValue.selection.end >= 4) selectionIndex++;
+    if (novoTextLength >= 5) {
+      newText.write(valorNovo.text.substring(2, usedSubstringIndex = 4) + '/');
+      if (valorNovo.selection.end >= 4) selectionIndex++;
     }
-    if (newTextLength >= usedSubstringIndex) {
-      newText.write(newValue.text.substring(usedSubstringIndex));
+    if (novoTextLength >= usedSubstringIndex) {
+      newText.write(valorNovo.text.substring(usedSubstringIndex));
     }
 
     return TextEditingValue(