+++ lib/src/formatters/hora_input_formatter.dart
@@ -7,43 +7,45 @@ class HoraInputFormatter extends TextInputFormatter {
 
   @override
   TextEditingValue formatEditUpdate(
-      TextEditingValue oldValue, TextEditingValue newValue) {
-    final newTextLength = newValue.text.length;
-    var selectionIndex = newValue.selection.end;
+      TextEditingValue valorAntigo, TextEditingValue valorNovo) {
+    final novoTextLength = valorNovo.text.length;
+    var selectionIndex = valorNovo.selection.end;
 
     var usedSubstringIndex = 0;
     final newText = StringBuffer();
 
-    if (newTextLength > maxLength) {
-      return oldValue;
+    if (novoTextLength > maxLength) {
+      return valorAntigo;
     }
 
-    switch (newTextLength) {
+    switch (novoTextLength) {
       case 1:
-        final hora = int.tryParse(newValue.text.substring(0, 1));
-        if (hora >= 3) return oldValue;
+        final hora = int.tryParse(valorNovo.text.substring(0, 1));
+        if (hora >= 3) return valorAntigo;
         break;
       case 2:
-        final hora = int.tryParse(newValue.text.substring(0, 2));
-        if (hora >= 24) return oldValue;
+        final hora = int.tryParse(valorNovo.text.substring(0, 2));
+        if (hora >= 24) return valorAntigo;
         break;
       case 3:
-        final minuto = int.tryParse(newValue.text.substring(2, 3));
-        if (minuto >= 6) return oldValue;
-        newText.write(newValue.text.substring(0, usedSubstringIndex = 2) + ':');
-        if (newValue.selection.end >= 2) selectionIndex++;
+        final minuto = int.tryParse(valorNovo.text.substring(2, 3));
+        if (minuto >= 6) return valorAntigo;
+        newText
+            .write(valorNovo.text.substring(0, usedSubstringIndex = 2) + ':');
+        if (valorNovo.selection.end >= 2) selectionIndex++;
         break;
       case 4:
-        final minuto = int.tryParse(newValue.text.substring(2, 4));
-        if (minuto >= 60) return oldValue;
-        newText.write(newValue.text.substring(0, usedSubstringIndex = 2) + ':');
-        if (newValue.selection.end >= 2) selectionIndex++;
+        final minuto = int.tryParse(valorNovo.text.substring(2, 4));
+        if (minuto >= 60) return valorAntigo;
+        newText
+            .write(valorNovo.text.substring(0, usedSubstringIndex = 2) + ':');
+        if (valorNovo.selection.end >= 2) selectionIndex++;
         break;
       default:
     }
 
-    if (newTextLength >= usedSubstringIndex) {
-      newText.write(newValue.text.substring(usedSubstringIndex));
+    if (novoTextLength >= usedSubstringIndex) {
+      newText.write(valorNovo.text.substring(usedSubstringIndex));
     }
 
     return TextEditingValue(