+++ lib/src/formatters/km_input_formatter.dart
@@ -7,40 +7,40 @@ class KmInputFormatter extends TextInputFormatter {
 
   @override
   TextEditingValue formatEditUpdate(
-      TextEditingValue oldValue, TextEditingValue newValue) {
-    final newTextLength = newValue.text.length;
-    var selectionIndex = newValue.selection.end;
+      TextEditingValue valorAntigo, TextEditingValue valorNovo) {
+    final novoTextLength = valorNovo.text.length;
+    var selectionIndex = valorNovo.selection.end;
 
-    if (newTextLength > maxLength) {
-      return oldValue;
+    if (novoTextLength > maxLength) {
+      return valorAntigo;
     }
 
     var usedSubstringIndex = 0;
     final newText = StringBuffer();
-    switch (newTextLength) {
+    switch (novoTextLength) {
       case 4:
-        newText.write(newValue.text.substring(0, usedSubstringIndex = 1) +
+        newText.write(valorNovo.text.substring(0, usedSubstringIndex = 1) +
             '.' +
-            newValue.text.substring(1, usedSubstringIndex = 3));
+            valorNovo.text.substring(1, usedSubstringIndex = 3));
         selectionIndex += 1;
         break;
       case 5:
-        newText.write(newValue.text.substring(0, usedSubstringIndex = 2) +
+        newText.write(valorNovo.text.substring(0, usedSubstringIndex = 2) +
             '.' +
-            newValue.text.substring(2, usedSubstringIndex = 4));
+            valorNovo.text.substring(2, usedSubstringIndex = 4));
         selectionIndex += 1;
         break;
       case 6:
-        newText.write(newValue.text.substring(0, usedSubstringIndex = 3) +
+        newText.write(valorNovo.text.substring(0, usedSubstringIndex = 3) +
             '.' +
-            newValue.text.substring(3, usedSubstringIndex = 5));
+            valorNovo.text.substring(3, usedSubstringIndex = 5));
         selectionIndex += 1;
         break;
       default:
     }
 
-    if (newTextLength >= usedSubstringIndex) {
-      newText.write(newValue.text.substring(usedSubstringIndex));
+    if (novoTextLength >= usedSubstringIndex) {
+      newText.write(valorNovo.text.substring(usedSubstringIndex));
     }
 
     return TextEditingValue(