+++ lib/src/formatters/validade_cartao_input_formatter.dart
@@ -7,35 +7,36 @@ class ValidadeCartaoInputFormatter extends TextInputFormatter {
 
   @override
   TextEditingValue formatEditUpdate(
-      TextEditingValue oldValue, TextEditingValue newValue) {
-    final newTextLength = newValue.text.length;
-    var selectionIndex = newValue.selection.end;
+      TextEditingValue valorAntigo, TextEditingValue valorNovo) {
+    final novoTextLength = valorNovo.text.length;
+    var selectionIndex = valorNovo.selection.end;
 
-    if (newTextLength > maxLength) {
-      return oldValue;
+    if (novoTextLength > maxLength) {
+      return valorAntigo;
     }
 
     var usedSubstringIndex = 0;
     final newText = StringBuffer();
 
-    switch (newTextLength) {
+    switch (novoTextLength) {
       case 1:
-        final hora = int.tryParse(newValue.text.substring(0, 1));
-        if (hora >= 2) return oldValue;
+        final hora = int.tryParse(valorNovo.text.substring(0, 1));
+        if (hora >= 2) return valorAntigo;
         break;
       case 2:
-        final hora = int.tryParse(newValue.text.substring(0, 2));
-        if (hora >= 13) return oldValue;
+        final hora = int.tryParse(valorNovo.text.substring(0, 2));
+        if (hora >= 13) return valorAntigo;
         break;
       case 3:
       case 4:
-        newText.write(newValue.text.substring(0, usedSubstringIndex = 2) + '/');
-        if (newValue.selection.end >= 2) selectionIndex++;
+        newText
+            .write(valorNovo.text.substring(0, usedSubstringIndex = 2) + '/');
+        if (valorNovo.selection.end >= 2) selectionIndex++;
         break;
     }
 
-    if (newTextLength >= usedSubstringIndex) {
-      newText.write(newValue.text.substring(usedSubstringIndex));
+    if (novoTextLength >= usedSubstringIndex) {
+      newText.write(valorNovo.text.substring(usedSubstringIndex));
     }
 
     return TextEditingValue(