+++ lib/src/modelos/regioes.dart
@@ -1,3 +1,7 @@
+/// Possui a lista com todas as regiões do Brasil.
+///
+/// `listaRegioes` lista com todas as regiões.
+
 class Regioes {
   static final List<String> listaRegioes = [
     'Centro-Oeste',