+++ lib/src/util/util_brasil_fields.dart
@@ -5,7 +5,8 @@ class UtilBrasilFields {
   }
 
   static String obterCpf(String cpf) {
-    assert(cpf.length == 11);
+    assert(cpf.length == 11,
+        'CPF com tamanho inválido. Deve conter 11 caracteres');
     return cpf.substring(0, 3) +
         '.' +
         cpf.substring(3, 6) +
@@ -16,7 +17,8 @@ class UtilBrasilFields {
   }
 
   static String obterCnpj(String cnpj) {
-    assert(cnpj.length == 14);
+    assert(cnpj.length == 14,
+        'CPNJ com tamanho inválido. Deve conter 14 caracteres');
     return cnpj.substring(0, 2) +
         '.' +
         cnpj.substring(2, 5) +
@@ -28,12 +30,35 @@ class UtilBrasilFields {
         cnpj.substring(12, 14);
   }
 
-  static String obterCep(String cep) {
-    assert(cep.length == 8);
+  static String obterCep(String cep, {bool ponto = true}) {
+    assert(
+        cep.length == 8, 'CEP com tamanho inválido. Deve conter 8 caracteres');
+    if (ponto) {
+      return cep.substring(0, 2) +
+          '.' +
+          cep.substring(2, 5) +
+          '-' +
+          cep.substring(5, 8);
+    }
     return cep.substring(0, 2) +
-        '.' +
         cep.substring(2, 5) +
         '-' +
         cep.substring(5, 8);
+  }
+
+  static String obterTelefone(String telefone, {bool ddd = true}) {
+    if (ddd) {
+      assert(telefone.length == 11,
+          'Telefone com tamanho inválido. Deve conter 11 caracteres');
+      return '(' +
+          telefone.substring(0, 2) +
+          ') ' +
+          telefone.substring(2, 7) +
+          '-' +
+          telefone.substring(7, 11);
+    }
+    assert(telefone.length == 9,
+        'Telefone com tamanho inválido. Deve conter 9 caracteres');
+    return telefone.substring(0, 5) + '-' + telefone.substring(5, 9);
   }
 }